ineedyou/ocr_test/requirements.txt
openai
requests
fastapi
uvicorn
python-dotenv
websockets